+++ sdk/program/src/wasm/mod.rs
@@ -3,7 +3,6 @@
 use wasm_bindgen::prelude::*;
 
 pub mod instructions;
-pub mod pubkey;
 pub mod system_instruction;
 
 /// Initialize Javascript logging and panic handler
